feat(api): expose generateClient from /server subpath (#12691)
* feat(api-graphql): implement server generateClient

* feat(api): expose generateClient from /server subpath

* chore(adapter-nextjs): adapter generateClient exported from api/server
