+++ packages/api-graphql/server/package.json
@@ -1,0 +1,7 @@
+{
+	"name": "@aws-amplify/api-graphql/server",
+	"types": "../dist/esm/server/index.d.ts",
+	"main": "../dist/cjs/server/index.js",
+	"module": "../dist/esm/server/index.mjs",
+	"sideEffects": false
+}
