+++ packages/api-graphql/src/internals/server/generateClientWithAmplifyInstance.ts
@@ -1,7 +1,7 @@
 // Copyright Amazon.com, Inc. or its affiliates. All Rights Reserved.
 // SPDX-License-Identifier: Apache-2.0
 
-import { graphql, cancel, isCancelError } from '../';
+import { graphql, cancel, isCancelError } from '..';
 import { generateModelsProperty } from './generateModelsProperty';
 import {
 	__amplify,
@@ -16,13 +16,15 @@ import {
 /**
  * @private
  *
+ * Used internally by `adapter-nextjs` package.
+ *
  * Creates a client that can be used to make GraphQL requests, using a provided `AmplifyClassV6`
  * compatible context object for config and auth fetching.
  *
  * @param params
  * @returns
  */
-export function generateClient<
+export function generateClientWithAmplifyInstance<
 	T extends Record<any, any> = never,
 	ClientType extends
 		| V6ClientSSRRequest<T>
