+++ packages/api/src/internals/index.ts
@@ -1,4 +1,4 @@
 // Copyright Amazon.com, Inc. or its affiliates. All Rights Reserved.
 // SPDX-License-Identifier: Apache-2.0
 export { InternalAPI, InternalAPIClass } from './InternalAPI';
-export { generateClient as generateServerClient } from '@aws-amplify/api-graphql/internals/server';
+export { generateClientWithAmplifyInstance } from '@aws-amplify/api-graphql/internals/server';
