+++ packages/api/src/server.ts
@@ -10,3 +10,4 @@ export {
 	patch,
 	isCancelError,
 } from '@aws-amplify/api-rest/server';
+export { generateClient } from '@aws-amplify/api-graphql/server';
